packages/infra_env/pyproject.toml
[project]
name = "infra-env"
version = "0.1.0"
description = "Utilidades de config de ambiente"
readme = "README.md"
authors = [
    { name = "gpupo", email = "g1mar@gpupo.com" }
]
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
